Consolidate functionality in NestedHostColumnVector and HostColumnVector (#6120)
