+++ java/src/test/java/ai/rapids/cudf/TableTest.java
@@ -187,18 +187,18 @@ public class TableTest extends CudfTestBase {
     }
   }
 
-  public static void assertPartialColumnsAreEqual(HostColumnVector.NestedHostColumnVector expected,
-                                                  HostColumnVector.NestedHostColumnVector cv,
+  public static void assertPartialColumnsAreEqual(HostColumnVectorCore expected,
+                                                  HostColumnVectorCore cv,
                                                   String colName, boolean enableNullCheck) {
     assertEquals(expected.getType(), cv.getType(), "Type For Column " + colName);
-    assertEquals(expected.getRows(), cv.getRows(), "Row Count For Column " + colName);
+    assertEquals(expected.getRowCount(), cv.getRowCount(), "Row Count For Column " + colName);
     if (enableNullCheck) {
       assertEquals(expected.getNullCount(), cv.getNullCount(), "Null Count For Column " + colName);
     } else {
       // TODO add in a proper check when null counts are supported by serializing a partitioned column
     }
     DType type = expected.getType();
-    for (int expectedRow = 0; expectedRow < expected.getRows(); expectedRow++) {
+    for (int expectedRow = 0; expectedRow < expected.getRowCount(); expectedRow++) {
       assertEquals(expected.isNull(expectedRow), cv.isNull(expectedRow),
           "NULL for Column " + colName + " Row " + expectedRow);
       if (!expected.isNull(expectedRow)) {
